New support for JSON config files
The biggest change is that all of the information that AbcSmc needs to run can
now be provided using a JSON config file.  There are probably things that need
to be supported that aren't yet, but it runs using the simple dice example.

This commit also fixes a methodological "bug": depending on the simulation, it
may be probable that one or more parameters will completely converge to a
single value.  In this case, the probability density at that value is infinite,
and some of the math in the method breaks down.  In that event, the code
previously would abort with an error about weights not being normalized (in
this case, because they're all infinite).  Now the code will instead continue
to run, sampling from particles uniformly with respect the parameters that have
converged to a single value.

Another minor bugfix: if the user specified a simulator that doesn't exist, now
the program fails gracefully, rather than repeatedly trying to run the
executable.

Finally, the output looks nicer, although it still could use some work.