Temporary bug fix for mpi STOP_TAG
The constant value we were using before appears to be too large an integer
for some hardware.  The way we're using message tags should be refactored,
probably so that there is one data tag and one stop tag, and the particle
id is prepended to the message rather than being encoded in the tag.  There
are no great solutions here--sending extra messages has overhead, changing
the data type makes things more complicated, and the other options involve
overloading a container that's in use for something else.  As long as we
reliably convert the particle id to a long double and back to the right
int, that approach will work fine.