Bugfix - All components were always being used
Desite the fact that we were calculating the optimal number of components to use
for each parameter, we were always using all components.  This would not have
caused a bias in estimated parameters, but it would have resulted in slower
convergence.

The plsr class function also had a redundant parameter that has been removed.