Resolved warning coming from jsoncpp
For reasons that are not clear, jsoncpp uses bitfields instead of bools. They
also were not being used correctly: a 1-bit int can only hold 0 or -1, rather
than 0 or 1, so it is not the right choice for a bool-replacement.  They have
been changed to unsigned ints, and the ridiculous, misleading compile warnings
are finally gone.