Workaround for Eigen assert crash when parameters converge
When one of the parameters converges completely on a single value, the z-scores
for that parameter are all undefined because you end up dividing by a 0 stdev.

Ideally, since that parameter cannot change or be correlated with other
parameters, it should be taken out of the PLS regression.  That is likely a
smarter solution, but a bit harder to implement, and this approach appears to
yield reasonable behavior in the unlikely event of complete convergence.