Grab 'R' particles to run if there are no more 'Q' particles
We are seeing that, in a set of 10,000 long simulation runs, 10 to 100
runs inexplicably take an insane amount of time, eventually timing out.
When repeated, they run without issues, using the same RNG seed.  In
order to deal with this ridiculousness automatically, a worker now will
grab a "running" (no longer running?) job if there are no more queued.