Trying to be cleverer about running jobs redundantly
Now tracking "attempts" in jobs table, so that particle coverage is more uniform.
This really only matterns in environments with a queueing system that will kill
off jobs after they run for some time.  Unfortunately, we've been experiencing
highly heterogeneous run times, apparently dependent on what else is going on on
a job's node.  This makes it sometimes worthwhile to run a simulation (with a
fixed seed) several times on different nodes, in the hopes that one at least
will finish in a sane amount of time.