Batch updates of jobs table
When simulations are fast, it makes sense to have threads do multiple
simulations per DB access, to avoid thrashing the I/O system.  We were
previously buffering reads and some writes to the DB, but not when
updating the posterior column in the jobs table.

Also, we recently discovered that there is some incompatibility between
optimization in gcc-5 and Eigen::eigensolver (it produces infinite values).
This problem does not occur with optimization (O2) and older versions of gcc,
or with optimization turned off (O0) under gcc-5.  I don't think I've ever
encountered an apparent bug in gcc before.  I don't know whether the bug is in
gcc-5, or if gcc-5 fixes a bug that Eigen was assuming.