Eigen/g++5 conflict resolved by switching to doubles
For reasons that aren't completely clear, when optimization is turned on and
g++ version >= 5, EigenSolver returns nans if the matrices contain long
doubles.  This bit of insanity does not come up if double are used.  Since that
is preferable to either avoiding optimization or being wedded to an old
compiler, that's the option we're going with for now.