Added support for more complicated parameter types
Parameter may now be modifiers of other parameters.  Modifications can be additive or
multiplicative, and may be applied in sampled space or transformed space.

As an example, we want to estimate a reporting rate for symptomatic infections.  We know
reporting practices changed in the middle of the modeled time period, however, so we want
to model a different historical rate.  Modeling these as two independent parameters would
imply that there is no relationship between the two, but as far as we know, things like
pathogenicity and health care seeking behavior have not changed, only reporting guidelines
for providers.

In this case, we have one parameter that represents a (modern) baseline for reporting
probability, and a second parameter that is normally-distributed in sampled (logit) space
that is added to the first.  Added values in logit space map to multiplied values in
untransformed (model) space.

NB: It is possible to create circular references, i.e. parameters that modify each other.
This is gibberish, and it's currently up to the user to ensure that this isn't specified.