Minor bugfix re: max psuedo parameter value
Using a non-int pseudo parameter internally represented as a double,
it was possible (but very uncommon) to get sampled parameter values
greater than the maximum the user requested.  This depended on the
step size and maximum used, and did not affect using the library for
doing ABC, only when doing sensitivity analyses/parameter sweeps.

Typically, the only effect would have been wasting compute time on
simulating some parameter combinations that were not of interest.